requirements.txt
streamlit
langchain
langchain-community
langchain-core
ollama
chromadb
faiss-cpu
unstructured
pytesseract
Pillow
gTTS